+++ libuna/libuna_base64_stream.c
@@ -1025,7 +1025,7 @@ int libuna_base64_stream_size_to_byte_stream(
 			}
 			if( character_limit != 0 )
 			{
-				if( number_of_characters != character_limit )
+				if( number_of_characters != (size_t) character_limit )
 				{
 					liberror_error_set(
 					 error,
@@ -1142,7 +1142,7 @@ int libuna_base64_stream_size_to_byte_stream(
 	}
 	if( character_limit != 0 )
 	{
-		if( number_of_characters != character_limit )
+		if( number_of_characters != (size_t) character_limit )
 		{
 			liberror_error_set(
 			 error,
@@ -1312,7 +1312,7 @@ int libuna_base64_stream_copy_to_byte_stream(
 			}
 			if( character_limit != 0 )
 			{
-				if( number_of_characters != character_limit )
+				if( number_of_characters != (size_t) character_limit )
 				{
 					liberror_error_set(
 					 error,
@@ -1425,7 +1425,7 @@ int libuna_base64_stream_copy_to_byte_stream(
 	}
 	if( character_limit != 0 )
 	{
-		if( number_of_characters != character_limit )
+		if( number_of_characters != (size_t) character_limit )
 		{
 			liberror_error_set(
 			 error,
@@ -1707,7 +1707,7 @@ int libuna_base64_stream_copy_from_byte_stream(
 		{
 			number_of_characters += 4;
 
-			if( number_of_characters >= character_limit )
+			if( number_of_characters >= (size_t) character_limit )
 			{
 				base64_stream[ base64_stream_index++ ] = (uint8_t) '\n';
